[ASTERIXDB-3649][*DB] Async request API backward compatibility fixes
- user model changes: no
- storage format changes: no
- interface changes: no

Ext-ref: MB-60882
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/20529
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ hyracks-fullstack/hyracks/hyracks-control/hyracks-control-cc/src/main/java/org/apache/hyracks/control/cc/result/IResultDirectoryService.java
@@ -40,7 +40,7 @@ public interface IResultDirectoryService extends IJobLifecycleListener, IResultM
             boolean emptyResult, int partition, int nPartitions, NetworkAddress networkAddress, String nodeId)
             throws HyracksDataException;
 
-    public void reportResultPartitionWriteCompletion(JobId jobId, ResultSetId rsId, int partition, int resultCount)
+    public void reportResultPartitionWriteCompletion(JobId jobId, ResultSetId rsId, int partition, long resultCount)
             throws HyracksDataException;
 
     public void reportResultPartitionConsumed(JobId jobId, ResultSetId rsId, int partition) throws HyracksDataException;
